vocabulo_junior_ml/requirements.txt
fastapi==0.112.2
Pillow==10.4.0
pillow_heif==0.18.0
python-dotenv==1.0.1
scikit_learn==1.3.0
spacy==3.7.2
SQLAlchemy==2.0.30
torch==2.4.0
transformers==4.41.2
uvicorn==0.30.6
